consumer contract tests `publish`
