+++ tests/README.md
@@ -5,5 +5,5 @@ This directory contains tests for `pubnub` crate.
 All tests can be executed with:
 
 ```shell
-cargo test
+cargo test --features contract_test --test contract_test
 ```
