+++ tests/contract_test.rs
@@ -15,6 +15,7 @@ async fn init_server(script: String) -> Result<String, Box<dyn std::error::Error
 async fn main() {
     env_logger::builder().try_init().unwrap();
     PubNubWorld::cucumber()
+        .max_concurrent_scenarios(1)
         .before(|_feature, _rule, scenario, _world| {
             futures::FutureExt::boxed(async move {
                 if scenario.tags.iter().any(|t| t.starts_with("contract=")) {
